+++ _templates/feature/content-source/npm-ignore.ejs.t
@@ -1,0 +1,8 @@
+---
+# helper funcs https://www.hygen.io/docs/templates/#helpers-and-inflections
+to: blocks/<%= h.inflection.dasherize(block_name) %>-block/sources/.npmignore
+---
+
+*.test.js
+*.test.jsx
+mock*.*