+++ _templates/feature/content-source/package.ejs.t
@@ -1,0 +1,6 @@
+---
+inject: true
+to: blocks/<%= h.inflection.dasherize(block_name) %>-block/package.json
+after: "files"
+---
+		"sources",