+++ _templates/feature/feature/storybook-main.ejs.t
@@ -1,0 +1,8 @@
+---
+inject: true
+to: blocks/<%= h.inflection.dasherize(block_name) %>-block/index.story.jsx
+append: true
+---
+export const ShowIcon = () => <<%= h.changeCase.pascal(feature_name) %> customFields={{ showIcon: true }} />;
+
+export const HideIcon = () => <<%= h.changeCase.pascal(feature_name) %> customFields={{ showIcon: false }} />;