+++ blocks/author-bio-block/index.story.jsx
@@ -242,3 +242,13 @@ export const notFoundSocialMedia = () => {
   };
   return (<AuthorBioItems content={data} />);
 };
+
+export const emptyImageObject = () => {
+  const { image, ...authorObjectWithoutImage } = authorObject;
+
+  const data = {
+    credits: { by: [authorObjectWithoutImage] },
+  };
+
+  return (<AuthorBioItems content={data} />);
+};