+++ blocks/identity-block/jest.config.js
@@ -2,4 +2,5 @@ const base = require('../../jest/jest.config.base');
 
 module.exports = {
   ...base,
+  testURL: 'http://localhost/',
 };