+++ jest/mocks/properties.js
@@ -1,1 +1,3 @@
-export default jest.fn();
+export default jest.fn(() => ({
+  locale: 'en',
+}));