+++ blocks/ads-block/CHANGELOG.md
@@ -3,6 +3,59 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.20.0](https://github.com/WPMedia/arc-themes-blocks/compare/@wpmedia/ads-block@5.19.0...@wpmedia/ads-block@5.20.0) (2022-01-26)
+
+
+### Bug Fixes
+
+* Freeze version to safe working one of arc ads per package uncertainty ([4f53caa](https://github.com/WPMedia/arc-themes-blocks/commit/4f53caabca3ee2d9cbcbfc893645d2d1fef61495))
+
+
+### Features
+
+* Update arcads to 6.1.1 ([120b78d](https://github.com/WPMedia/arc-themes-blocks/commit/120b78ddebdf38bb0ca05f50c67dda1c6bf9c39e))
+
+
+### Reverts
+
+* Revert "chore: Update arcads to 6 for ad block" ([e9b803e](https://github.com/WPMedia/arc-themes-blocks/commit/e9b803e7e6b05514f56701bd36142b2024668b7f))
+
+
+
+## 5.11.1-hotfix.0 (2021-04-23)
+
+
+
+## 5.10.3-hotfix.0 (2021-04-12)
+
+
+### Bug Fixes
+
+* Match default value to a designated block ([991238b](https://github.com/WPMedia/arc-themes-blocks/commit/991238b273c59c092090adffc1b88133b287a812))
+
+
+
+# 5.10.0 (2021-02-17)
+
+
+
+## 5.9.4 (2021-01-14)
+
+
+### Bug Fixes
+
+* Remove exhaustive hooks issue and remove unnecessary eslint ignore ([8bed4ab](https://github.com/WPMedia/arc-themes-blocks/commit/8bed4ab8524ff27814023ca85af2613307a2510d))
+
+
+### Performance Improvements
+
+* Remove mock feature files ([bdd5a3b](https://github.com/WPMedia/arc-themes-blocks/commit/bdd5a3bc942ac93a97623bf5c1fdd3aec264aa6f))
+* Remove testfiles from package with nested npmignore to override files declaration ([7af4ec9](https://github.com/WPMedia/arc-themes-blocks/commit/7af4ec9b0973341302934d001886bd68bf64ad78))
+
+
+
+
+
 # [5.19.0](https://github.com/WPMedia/arc-themes-blocks/compare/@wpmedia/ads-block@5.18.2...@wpmedia/ads-block@5.19.0) (2021-12-15)
 
 **Note:** Version bump only for package @wpmedia/ads-block