+++ blocks/resizer-image-block/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.18.0](https://github.com/WPMedia/arc-themes-blocks/compare/@wpmedia/resizer-image-block@5.17.0...@wpmedia/resizer-image-block@5.18.0) (2022-01-26)
+
+**Note:** Version bump only for package @wpmedia/resizer-image-block
+
+
+
+
+
 # [5.17.0](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/resizer-image-block@5.16.0...@wpmedia/resizer-image-block@5.17.0) (2021-11-17)